+++ test/trim/Trimmability/runtests.jl
@@ -19,4 +19,5 @@ outdir = ARGS[1]
     @test lines[9] == "finalizers: 27 32"
     @test lines[10] == "collected: 0 kept, 10 dropped"
     @test lines[11] == "got 1:3 vs 1 with 2.5 and sym and c"
+    @test lines[12] == "custom_library_ccall: $(VERSION.major)"
 end